requirements.txt
fastapi
uvicorn
langchain
langchain-core
langchain-community
langchain-groq
python-dotenv
requests
